DEMRUM-5771: Fix lint & format
